Add Python script to split GAP data to single matches.
The GAP data containss of 4000 rows with the following columns:
ID	Text	Pronoun	Pronoun-offset	A	A-offset	A-coref	B	B-offset	B-coref	URL

This script splits each row into two with the columns:
ID	Text	Pronoun	Pronoun-offset	Name	Name-offset	Coref	URL

The first row drops the B columns and the second row the A columns.
The columns ID, Text, Pronoun, Pronoun-Offset, and URL appear unchanged in both rows.
The remaining A and B columns are renamed to Name, Name-Offset, and Coref.

The first row uses the A columns.